chore: Updates wasi-logging to have a version
As I was starting to build and push all of these interfaces to work with
wkg, I realized this didn't even have a version yet which means `wkg`
won't even work with it. It also should just have a version. I am just
making this a plain 0.1.0 because versions are free and they don't need
to match the other p2 worlds. If we want to modify this, we just bump to
a 0.2.0-beta or something similar

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>